+++ loogo/loogo_test.go
@@ -1,0 +1,3 @@
+package loogo
+
+import _ "reflect"
